+++ server/src/main/java/org/opensearch/search/profile/ProfileResult.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.search.profile;
 
+import org.opensearch.Version;
 import org.opensearch.core.ParseField;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
@@ -44,8 +45,10 @@ import org.opensearch.core.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.LinkedHashMap;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
@@ -69,7 +72,13 @@ public final class ProfileResult implements Writeable, ToXContentObject {
     static final ParseField BREAKDOWN = new ParseField("breakdown");
     static final ParseField DEBUG = new ParseField("debug");
     static final ParseField NODE_TIME = new ParseField("time");
+    static final ParseField MAX_SLICE_NODE_TIME = new ParseField("max_slice_time");
+    static final ParseField MIN_SLICE_NODE_TIME = new ParseField("min_slice_time");
+    static final ParseField AVG_SLICE_NODE_TIME = new ParseField("avg_slice_time");
     static final ParseField NODE_TIME_RAW = new ParseField("time_in_nanos");
+    static final ParseField MAX_SLICE_NODE_TIME_RAW = new ParseField("max_slice_time_in_nanos");
+    static final ParseField MIN_SLICE_NODE_TIME_RAW = new ParseField("min_slice_time_in_nanos");
+    static final ParseField AVG_SLICE_NODE_TIME_RAW = new ParseField("avg_slice_time_in_nanos");
     static final ParseField CHILDREN = new ParseField("children");
 
     private final String type;
@@ -77,7 +86,11 @@ public final class ProfileResult implements Writeable, ToXContentObject {
     private final Map<String, Long> breakdown;
     private final Map<String, Object> debug;
     private final long nodeTime;
+    private final long maxSliceNodeTime;
+    private final long minSliceNodeTime;
+    private final long avgSliceNodeTime;
     private final List<ProfileResult> children;
+    private final boolean concurrent;
 
     public ProfileResult(
         String type,
@@ -87,12 +100,31 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         long nodeTime,
         List<ProfileResult> children
     ) {
+        this(type, description, breakdown, debug, nodeTime, children, false, -1, -1, -1);
+    }
+
+    public ProfileResult(
+        String type,
+        String description,
+        Map<String, Long> breakdown,
+        Map<String, Object> debug,
+        long nodeTime,
+        List<ProfileResult> children,
+        boolean concurrent,
+        long maxSliceNodeTime,
+        long minSliceNodeTime,
+        long avgSliceNodeTime
+    ) {
         this.type = type;
         this.description = description;
         this.breakdown = Objects.requireNonNull(breakdown, "required breakdown argument missing");
         this.debug = debug == null ? Map.of() : debug;
         this.children = children == null ? List.of() : children;
         this.nodeTime = nodeTime;
+        this.concurrent = concurrent;
+        this.maxSliceNodeTime = maxSliceNodeTime;
+        this.minSliceNodeTime = minSliceNodeTime;
+        this.avgSliceNodeTime = avgSliceNodeTime;
     }
 
     /**
@@ -105,6 +137,23 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         breakdown = in.readMap(StreamInput::readString, StreamInput::readLong);
         debug = in.readMap(StreamInput::readString, StreamInput::readGenericValue);
         children = in.readList(ProfileResult::new);
+        if (in.getVersion().onOrAfter(Version.V_3_0_0)) {
+            this.concurrent = in.readBoolean();
+            if (concurrent) {
+                this.maxSliceNodeTime = in.readLong();
+                this.minSliceNodeTime = in.readLong();
+                this.avgSliceNodeTime = in.readLong();
+            } else {
+                this.maxSliceNodeTime = -1;
+                this.minSliceNodeTime = -1;
+                this.avgSliceNodeTime = -1;
+            }
+        } else {
+            this.concurrent = false;
+            this.maxSliceNodeTime = -1;
+            this.minSliceNodeTime = -1;
+            this.avgSliceNodeTime = -1;
+        }
     }
 
     @Override
@@ -115,6 +164,14 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         out.writeMap(breakdown, StreamOutput::writeString, StreamOutput::writeLong);
         out.writeMap(debug, StreamOutput::writeString, StreamOutput::writeGenericValue);
         out.writeList(children);
+        if (out.getVersion().onOrAfter(Version.V_3_0_0)) {
+            out.writeBoolean(concurrent);
+            if (concurrent) {
+                out.writeLong(maxSliceNodeTime);
+                out.writeLong(minSliceNodeTime);
+                out.writeLong(avgSliceNodeTime);
+            }
+        }
     }
 
     /**
@@ -154,6 +211,22 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         return nodeTime;
     }
 
+    public long getMaxSliceTime() {
+        return maxSliceNodeTime;
+    }
+
+    public long getMinSliceTime() {
+        return minSliceNodeTime;
+    }
+
+    public long getAvgSliceTime() {
+        return avgSliceNodeTime;
+    }
+
+    public boolean isConcurrent() {
+        return concurrent;
+    }
+
     /**
      * Returns a list of all profiled children queries
      */
@@ -168,9 +241,19 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         builder.field(DESCRIPTION.getPreferredName(), description);
         if (builder.humanReadable()) {
             builder.field(NODE_TIME.getPreferredName(), new TimeValue(getTime(), TimeUnit.NANOSECONDS).toString());
+            if (concurrent) {
+                builder.field(MAX_SLICE_NODE_TIME.getPreferredName(), new TimeValue(getMaxSliceTime(), TimeUnit.NANOSECONDS).toString());
+                builder.field(MIN_SLICE_NODE_TIME.getPreferredName(), new TimeValue(getMinSliceTime(), TimeUnit.NANOSECONDS).toString());
+                builder.field(AVG_SLICE_NODE_TIME.getPreferredName(), new TimeValue(getAvgSliceTime(), TimeUnit.NANOSECONDS).toString());
+            }
         }
         builder.field(NODE_TIME_RAW.getPreferredName(), getTime());
-        builder.field(BREAKDOWN.getPreferredName(), breakdown);
+        if (concurrent) {
+            builder.field(MAX_SLICE_NODE_TIME_RAW.getPreferredName(), getMaxSliceTime());
+            builder.field(MIN_SLICE_NODE_TIME_RAW.getPreferredName(), getMinSliceTime());
+            builder.field(AVG_SLICE_NODE_TIME_RAW.getPreferredName(), getAvgSliceTime());
+        }
+        createBreakownView(builder);
         if (false == debug.isEmpty()) {
             builder.field(DEBUG.getPreferredName(), debug);
         }
@@ -184,6 +267,24 @@ public final class ProfileResult implements Writeable, ToXContentObject {
         }
 
         return builder.endObject();
+    }
+
+    private void createBreakownView(XContentBuilder builder) throws IOException {
+        Map<String, Long> modifiedBreakdown = new LinkedHashMap<>(breakdown);
+        if (!concurrent) {
+            removeStartTimeFields(modifiedBreakdown);
+        }
+        builder.field(BREAKDOWN.getPreferredName(), modifiedBreakdown);
+    }
+
+    static void removeStartTimeFields(Map<String, Long> modifiedBreakdown) {
+        Iterator<Map.Entry<String, Long>> iterator = modifiedBreakdown.entrySet().iterator();
+        while (iterator.hasNext()) {
+            Map.Entry<String, Long> entry = iterator.next();
+            if (entry.getKey().endsWith("_startTime")) {
+                iterator.remove();
+            }
+        }
     }
 
     private static final InstantiatingObjectParser<ProfileResult, Void> PARSER;